fix: Slider 컴포넌트로 분리
- const [sliderPosition, setSliderPosition] = useState(0);이거 꼭 써야하나
