+++ src/ui/ListView.jsx
@@ -1,7 +1,7 @@
-import { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 import styled, { keyframes, css } from 'styled-components';
 import { ListViewItem } from './ListViewItem';
-import { RightOutlined, LeftOutlined } from '@ant-design/icons';
+import { Slider } from './Slider';
 
 export function ListView({
 	newsData,
@@ -11,7 +11,7 @@ export function ListView({
 }) {
 	const [isActive, setIsActive] = useState(0);
 	const [indicator, setIndicator] = useState(1);
-
+	const [sliderPosition, setSliderPosition] = useState(0);
 	const categorizedData = Object.values(newsData).reduce((acc, cur) => {
 		const category = cur.category;
 		if (!acc[category]) {
@@ -21,7 +21,7 @@ export function ListView({
 		acc[category].push(data);
 		return acc;
 	}, {});
-	// console.log('categorizedData', categorizedData);
+
 	// 카테고리 기준으로 데이터 가공
 	const categoryList = Object.entries(categorizedData);
 
@@ -29,8 +29,6 @@ export function ListView({
 		setIsActive(index);
 		setSliderPosition(categoryStartIndex[index]);
 	}
-	//이미지 슬라이드
-	const [sliderPosition, setSliderPosition] = useState(0);
 
 	const { totalSlides, categoryStartIndex } = categoryList.reduce(
 		({ totalSlides, categoryStartIndex }, [_, items]) => {
@@ -41,13 +39,10 @@ export function ListView({
 		{ totalSlides: 0, categoryStartIndex: [] }
 	);
 
-	// 슬라이더 이동 로직
-	const moveSlider = direction => {
-		if (direction === 'left' && sliderPosition > 0) {
-			setSliderPosition(sliderPosition - 1);
+	const tabIndicator = direction => {
+		if (direction === 'left') {
 			setIndicator(indicator - 1);
-		} else if (direction === 'right' && sliderPosition < totalSlides - 1) {
-			setSliderPosition(sliderPosition + 1);
+		} else if (direction === 'right') {
 			setIndicator(indicator + 1);
 		}
 	};
@@ -73,60 +68,36 @@ export function ListView({
 						</StyledCategoryTab>
 					))}
 				</StyledCatetoryList>
-				<StyledListViewItem>
-					<StyledDiv
-						style={{
-							transform: `translateX(-${sliderPosition * 100}%)`, // 슬라이더 이동
-							transition: 'all 0.5s',
-						}}
-					>
-						{Object.values(categorizedData).map((item, index) => (
-							<ListViewItem
-								key={index}
-								categorizedData={item}
-								fetchSubscriptionData={fetchSubscriptionData}
-								setPopup={setPopup}
-								setSelectedPress={setSelectedPress}
-							/>
-						))}
-					</StyledDiv>
-					<StyledButton
-						as={LeftOutlined}
-						onClick={() => moveSlider('left')}
-						disabled={sliderPosition === 0}
-					/>
-
-					<StyledButton
-						as={RightOutlined}
-						onClick={() => {
-							moveSlider('right');
-						}}
-						disabled=''
-					/>
+				<StyledListViewItem
+					totalSlides={totalSlides}
+					categoryStartIndex={categoryStartIndex}
+					tabIndicator={tabIndicator}
+					sliderPosition={sliderPosition}
+					setSliderPosition={setSliderPosition}
+				>
+					{Object.values(categorizedData).map((item, index) => (
+						<ListViewItem
+							key={index}
+							categorizedData={item}
+							fetchSubscriptionData={fetchSubscriptionData}
+							setPopup={setPopup}
+							setSelectedPress={setSelectedPress}
+						/>
+					))}
 				</StyledListViewItem>
 			</StyledWrapper>
 		</>
 	);
 }
-const StyledDiv = styled.div`
-	position: relative;
-	width: 100%;
-	height: 100%;
-	display: flex;
-	flex: 0 0 100%;
-`;
+
 const StyledWrapper = styled.div`
 	position: relative;
 	width: 100%;
 	height: 388px;
 	padding: 0;
 `;
-const StyledListViewItem = styled.div`
-	display: flex;
-	width: 100%;
-	height: 100%;
-	overflow: hidden;
-`;
+const StyledListViewItem = styled(Slider)``;
+
 const StyledCatetoryList = styled.div`
 	display: flex;
 	align-items: center;
@@ -167,23 +138,6 @@ const StyledCategoryTab = styled.button`
 				animation: ${AnimationFill} 20s forwards;
 			`}
     }
-`;
-const StyledButton = styled.button`
-	position: absolute;
-	top: 50%;
-	left: -48px;
-	transfor: translateY(-50%);
-	font-size: 42px;
-	padding: 20px 0;
-	cursor: pointer;
-	&:disabled {
-		background-color: #ddd;
-		display: none;
-	}
-	&:last-child {
-		left: auto;
-		right: -48px;
-	}
 `;
 
 const AnimationFill = keyframes`
