--- src/ui/test.jsx
@@ -1,7 +1,0 @@
-export function testComp() {
-	return (
-		<>
-			<h1>Test Component</h1>
-		</>
-	);
-}
