--- src/utility/swiper.js
@@ -1,12 +1,0 @@
-/**
- * 공통
- * 스와이퍼 필요한 기능
- * 버튼 클릭시 이동 (이전, 다음)
- * 제일 처음은 이전 버튼 hidden
- * 제일 마지막은 다음 버튼 hidden
- * 리스트형태 경우
- * 탭과 스와이퍼 연동
- * 탭 클릭하면 해당 스와이퍼로 이동
- */
-
-export function swiperController() {}
